% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_allen_geomx.R
\name{merge_allen_geomx}
\alias{merge_allen_geomx}
\title{Merge Seurat Object with Allen SEA AD Atlas}
\usage{
merge_allen_geomx(seurat_object, allen_object)
}
\arguments{
\item{seurat_object}{A Seurat object to be merged.}

\item{allen_object}{A Seurat object representing the Allen SEA AD Atlas.}
}
\value{
A merged Seurat object containing only the common features between the two datasets.
}
\description{
This function merges a Seurat object with the Allen SEA AD Atlas after finding and subsetting
both datasets to their common features (row names).
}
\examples{
# Assuming seurat_object and Allen_SEA_AD_Atlas are pre-loaded Seurat objects
merged_seurat <- merge_allen_geomx(seurat_object, Allen_SEA_AD_Atlas)
}
