% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_seurat_objects.R
\name{merge_seurat_objects}
\alias{merge_seurat_objects}
\title{Merge Two Seurat Objects}
\usage{
merge_seurat_objects(seurat_object, Allen_SEA_AD_Atlas)
}
\arguments{
\item{seurat_object}{A Seurat object to be merged.}

\item{Allen_SEA_AD_Atlas}{Another Seurat object to merge with `seurat_object`.}
}
\value{
A new Seurat object containing the merged counts and metadata.
}
\description{
This function merges two Seurat objects by combining their count matrices and metadata.
It ensures compatibility between the two datasets and creates a new Seurat object.
}
\examples{
# Example usage:
seurat_merge <- merge_seurat_objects(seurat_object, Allen_SEA_AD_Atlas)
}
