% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top20_enriched_pathways.R
\name{plot_top20_enriched_pathways}
\alias{plot_top20_enriched_pathways}
\title{Identify and Plot Top 20 Enriched Pathways for Correlated NMF Factors}
\usage{
plot_top20_enriched_pathways(
  seurat_object,
  allen_object,
  nmf_factor = 9,
  output_file = "top20.pdf"
)
}
\arguments{
\item{seurat_object}{A Seurat object containing GSEA results in the NMF reduction for the Seurat dataset.}

\item{allen_object}{A Seurat object containing GSEA results in the NMF reduction for the Allen dataset.}

\item{nmf_factor}{The NMF factor to analyze for enrichment (default is 9).}

\item{output_file}{The path to save the output bar chart as a PDF (default is "top20.pdf").}
}
\value{
A ggplot object of the bar chart for the top 20 enriched pathways.
}
\description{
This function identifies the top 20 gene sets enriched in the most correlated NMF factors
for spatial and single-cell datasets and generates a comparative bar chart.
}
\examples{
library(Seurat)
library(ggplot2)
p <- plot_top20_enriched_pathways(
  seurat_object = seurat_object,
  allen_object = allen_object,
  nmf_factor = 9,
  output_file = "top20_enriched.pdf"
)
print(p)
}
