% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top20_enriched_pathways_single.R
\name{plot_top20_enriched_pathways_single}
\alias{plot_top20_enriched_pathways_single}
\title{Identify and Plot Top 20 Enriched Pathways for Correlated NMF Factor}
\usage{
plot_top20_enriched_pathways_single(
  seurat_object,
  nmf_factor = 9,
  output_file = "top20_dotplot.pdf"
)
}
\arguments{
\item{seurat_object}{A Seurat object containing GSEA results in the NMF reduction.}

\item{nmf_factor}{The NMF factor to analyze for enrichment (default is 9).}

\item{output_file}{The path to save the output dot plot as a PDF (default is "top20_dotplot.pdf").}
}
\value{
A ggplot object of the dot plot for the top 20 enriched pathways.
}
\description{
This function identifies the top 20 gene sets enriched in the specified NMF factor
for a given dataset, weighted by NES and p-values, and generates a dot plot
of the top 20 enriched pathways.
}
\examples{
library(Seurat)
library(ggplot2)
p <- plot_top20_enriched_pathways_dotplot(
  seurat_object = seurat_object,
  nmf_factor = 9,
  output_file = "top20_dotplot.pdf"
)
print(p)
}
