% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_run_nmf.R
\name{process_and_run_nmf}
\alias{process_and_run_nmf}
\title{Perform Preprocessing, PCA, and NMF on a Seurat Object}
\usage{
process_and_run_nmf(
  seurat_object,
  nmf_rank = NULL,
  nmf_reduction_name = "nmf",
  nmf_reduction_key = "NMF_",
  seed = 16
)
}
\arguments{
\item{seurat_object}{A Seurat object to be processed.}

\item{nmf_rank}{The number of factors (k) for NMF. If NULL, automatic rank determination is performed.}

\item{nmf_reduction_name}{The name of the reduction slot to assign the NMF results. Default is "nmf".}

\item{nmf_reduction_key}{The prefix for NMF dimensions in the reduction slot. Default is "NMF_".}

\item{seed}{The random seed for reproducibility. Default is 16.}
}
\value{
A processed Seurat object with PCA and NMF results.
}
\description{
This function performs data normalization, scaling, PCA, and NMF on a Seurat object.
It assigns the results to the appropriate slots and provides tools for visualization.
}
\examples{
library(Seurat)
seurat_object <- CreateSeuratObject(counts = matrix(rnorm(1000), nrow = 10))
seurat_object <- process_and_run_nmf(seurat_object, nmf_rank = 5)
RankPlot(seurat_object, detail.level = 2)
}
